break canTransform loop for negative time delta only when simulated
(At least) with Python 2 ros.Time.now() is not necessarily monotonic
and one can experience negative time deltas (<1s) on real hardware under full load.

This check was originally introduced to allow for backjumps with rosbag replays,
and only there it makes sense. So we'll check for backjumps only when use_sim_time
is set.